requirements.txt
fastapi==0.115.6
pydantic==2.10.3
python-dotenv==1.0.1
tensorflow==2.18.0
uvicorn==0.32.1
